refactor: move ExpandingContainer to ArchivedComponents
